Added Decorations next to Category Display +
The displayed category text is now in a horizontal box, so the decorations on the left and right will be automatically aligned with the text, regardless of the font size.

If there are no items in the current category, NoItemsTextBlock is displayed.

Once the player selects a category, the button cannot be clicked again.